keep: warningレベルを上げた際に発生するwarningの一部を修正 以下のwarningを修正しました 　warning C4244: 'xx' : 'yyyy' から 'zzzz' への変換です。データが失われる可能性があります。 　warning C4310: キャストによって定数値が切り捨てられました。 　　(ANSI構成での1箇所対応しない箇所が有ります) 　warning C4701: 初期化されていない可能性のあるローカル変数 'xxxx' が使用されます 　warning C4702: 制御が渡らないコードです。 VS2010Expressで評価しました Patchunicode#372
git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk2@2745 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/mem/CNativeA.cpp
@@ -350,7 +350,7 @@ void CNativeA::ToZenkaku(
 					usDes = (unsigned short)0x8394; /* ƒ” */
 					nCharChars = 2;
 				}else {
-					usDes = _mbbtombc( usSrc );
+					usDes = (unsigned short)_mbbtombc( usSrc );
 					/* ‘÷‰¹ */
 					if( bHiragana != 2 && pBuf[i + 1] == (unsigned char)'Þ' && NULL != strchr( (const char *)pszDakuSet, pBuf[i] ) ){
 						usDes++;
